+++ jest.config.js
@@ -26,7 +26,7 @@ module.exports = {
 
   // The glob patterns Jest uses to detect test files
   testMatch: ["**/test/unit/**/?(*.)+(spec|test).js?(x)"],
-  // testMatch: ["**/test/unit/Orchestrator.spec.js"],
+  // testMatch: ["**/test/unit/plugins/Plugins.spec.js"],
 
   // The test environment that will be used for testing
   testEnvironment: "node"